Add `read_only` field setting
If present and set to true, the field will be rendered on forms but
with a 'readonly' attribute set. It will also be greyed out. This is
intended for use with fields that should be visible to users, but not
updated. An example would be fields storing IDs that link entities to
external systems. It would be useful for users to be able to see these,
but not update them as this could cause inconsistencies.

Once field-level permissions have been added, this feature may become
redundant. Fields that the user has permission to view but not edit
could automatically be disabled and greyed out.

An example field definition:
```
{
  "title": "External ID",
  "identifier": "external_id",
  "type": "text",
  "settings": {
    "read_only": true
  }
}
```