resolves #407 - Add `hidden` field setting
If this setting is present and set to true, the field will be hidden
on forms. One use case is for a sort index field linked to a
drag-and-drop list interface. Dragging list items to change their
order will update the linked field, but we don't want the actual
position field visible to users.

An example field definition:
```
{
  "title": "Position",
  "identifier": "position",
  "type": "sortindex",
  "settings": {
    "hidden": true
  }
}
```