resolves #407 - Add `hidden` field setting
If present and set to true, the field will not be rendered on forms.
The field will remain editable by the user, it just won't be displayed.
One use case is for a sort index field linked to a drag-and-drop list
interface. Dragging list items to change their order will update the
linked field, but we don't want the actual position field visible to
users.

An example field definition:
```
{
  "title": "Position",
  "identifier": "position",
  "type": "sortindex",
  "settings": {
    "hidden": true
  }
}
```