Add `max_displayed_rows` setting for table collection fields
If present, at most this number of collection items will be displayed
in each table row.

One use case for this is only displaying the first product image in a
product list view. Products can have more than one image, but you
wouldn't normally want to show all images for each product in a table
row. With this setting, we can only show the first image.

An example field definition:
```
{
    "title": "View with Image thumbnail",
    "identifier": "with_thumbnail",
    "type": "table",
    "settings": {
        "sort": {
            "field": "created",
            "asc": false,
            "sortable": false
        },
        "fields": {
            "name": {
                "label": "Name"
            },
            "images": {
                "label": "Image Previews",
                "type": "collection",
                "settings": {
                    "fields": {
                        "url": {
                            "label": "Preview"
                        }
                    },
                    "max_displayed_rows": 1
                }
            }
        }
    }
}
```