Tree view sorting fix #2: new item positions
New items previously got a position equal to the number of items of
that content type. In a flat list, this is fine - items get added to
the end of the list. In a tree, though, this means that new items
added to lower levels of the tree could get high position numbers,
leading to gaps in the position sequence.

New items now get a position value equal to the number of items that
share a content type and parent value, so they'll get sorted to the
end of their tree siblings.