Tree view sorting fix #3: deleting items
When an item is deleted, all other items of that content type with a
higher position have their position decremented. The issue is that
this applies to every tree item, even ones on different branches or
levels.

This commit changes the soft delete method to only modify positions
on sibling items.