Tree view sorting fix #4: changing parents
If an item is moved from under one parent to a different one, then:

1. Any of its previous siblings with higher positions need to be moved
   down, otherwise there'll be a gap in the numbering.
2. If an item is moved to the same position under a different parent,
   its new siblings need to be shifted up to make room.

At the moment, the situation where an item changes position but stays
under the same parent is effectively handled as a special case. It
might be possible to simplify the logic to only ever do two shifts
(shift down from where the item's removed, shift up from where the
item is moved to), but this would be less efficient on large
collections.