Tree view sorting fix #5: moving to root
Previously, moving a tree item to the root from a lower level didn't
work. This was because the `parent` field was sent as null, which
meant the Symfony form builder ignored the `parent` field. (See
vendor/symfony/symfony/src/Symfony/Component/Form/Form.php::submit
for the relevant logic - seems to be because parent is a composite
field and $clearMissing is set to false.)

To resolve this, the tree view JS now checks if `parent` is null in
the updateRow function, and if so replaces the value with
`{ content: null }`. This sets the `content` sub-field to null, which
allows the item to be moved to the root level.