fix #513: Missing select icons in reference modal table
In some cases, the table in the modal lightbox used to select an item
for a reference field would be missing the circular select icons. This
appeared to be due to the inline `width` style set on the table header
and body cell divs. This was either set to `0px` or a non-zero value.
If it was `0px`, the selection icon displayed correctly. If not, the
selection icon wouldn't be visible.

The inline width style is added by a Vue update lifecycle hook in
`TableContent.vue`. Changing this to always set the width to `0px`
appears to fix the issue.