Fix duplicate reference selections
When a selection modal for a reference field is shown, an observer on
the `contentSelected` event is added. When an item is selected, the
observer is removed and the lightbox is closed. However, if the
lightbox is closed some other way (e.g. by hitting escape or by
clicking outside the modal), the observer won't be removed. This means
selections made in one modal can affect others. To replicate:

1. Add blank reference field rows to a collection field until there
   are at least two empty rows.
2. Click on a blank row to bring up the modal.
3. Close the modal without selecting anything (hit escape).
4. Click on another row to open its modal, and select an item.
5. Both rows will now update to show the selected item.

To resolve this issue, the observer is instead removed on the UIkit
modal's `beforeHide` event. This means the observer will be removed no
matter how the modal is closed.

The `closeModal` method in the `field/Reference.vue` component was
only used within this file, and was removed as it's no longer
necessary.