Add new ColourPicker field type
This feature allows selecting a colour using the Vue component vue-color. After selecting the input the picker dropdown will be displayed which allows selection of the colour's hue, saturation and alpha values.

Alternatively the hex code or rgba value can be entered, which is persisted to the database as a string.

For more information about `vue-color`, see http://vue-color.surge.sh