Constrain modal header width on screens wider than 2100px
This fix sets a max-width of 100% to prevent the header of a view modal
from overflowing horizontally.

This improves the look of the modals as it ensures the close icon is
always positioned in the top right corner of the modal, as well as the
search input always being visible.