Validate not_empty setting on reference fields
The `not_empty` setting doesn't work for reference fields, as the
domain and content type values are always present even if the content
value is blank. These two values are enough to allow the `not_empty`
validation in UniteCMSCoreTypeExtension::buildForm to pass, even if no
content value has been provided.

This commit modifies the validateData method of the reference field
type to not skip validation when $data is empty if `not_empty` is set.
The existing validation code will then trigger an error if any of
domain/content_type/content are missing.

This commit also modifies the error message shown to be the standard
"This field is required", rather than a more complex message asking
the use to contact an administrator. A reference field can't be added
to the domain without specifying domain and content_type, so the only
reason validation should fail is because content is missing. The
simpler error message should be less confusing for less technical
users.