Add min, max & step settings for Date and DateTime fields
This feature adds support for the attributes `min`, `max` and `step` in
the field types `date` and `datetime`. These attributes are for the HTML5
input types `date` and `datetime-local`, which include validation to
ensure the attributes are successfully applied to the input elements.

An example of definition of these fields in the domain are as follows:

```
{
  "title": "Date Posted",
  "identifier": "date_posted",
  "type": "date",
  "settings": {
    "description": "",
    "min": "2019-01-01",
    "max": "2019-03-03",
    "step": 2
  }
},
{
  "title": "Date Time Posted",
  "identifier": "datetime_posted",
  "type": "datetime",
  "settings": {
    "description": "",
    "min": "2019-02-02T00:00",
    "max": "2019-03-03T12:59",
    "step": 2
  }
}
```