Diff view for content revisions. This closes #669
This feature allows viewing the diffs between two content revisions
using the ace editor. It has the following main features:

1. Add the boolean setting 'revision_description' to text and textarea
fields. If the field is defined, when displaying the revision overview,
the value of this field will be used in the 'description' column of the
table view. If there is no 'revision_description' field defined, the
column will not be rendered. The description text will also only be
displayed on the diff page if it is defined.

3. In the content revisions interface, don't show the content for each
revision. Instead, just show the version number, description, action,
timestamp and actor fields in a tabular view.

4. Add an interface to allow selecting a range of revisions: show two
radio button columns in the content revisions tabular page. In order to
allow viewing just a single revision, you should be able to select the
same radio button in both columns.

5. Once a revision range is selected, clicking button 'View Diff' lets
you view a diff between the two versions. If a single revision is
selected in both columns, you'd just show that rather than a diff.